Upgrade to `marshmallow-dataclass` 8
This is done in order to be able to use `base_schema` in `class_schema`
In order for this to work we need to move some classes around to avoid
cyclic dependencies.
