+++ requirements/requirements-dev.txt
@@ -81,7 +81,7 @@ jsonschema==3.2.0         # via -r requirements.txt, matrix-synapse, web3
 lazy-object-proxy==1.4.1  # via astroid
 lru-dict==1.1.6           # via -r requirements.txt, web3
 markupsafe==1.1.1         # via -r requirements-docs.txt, -r requirements.txt, jinja2
-marshmallow-dataclass==6.0.0  # via -r requirements.txt
+marshmallow-dataclass==8.0.0  # via -r requirements.txt
 marshmallow-enum==1.5.1   # via -r requirements.txt
 marshmallow-polyfield==5.9  # via -r requirements.txt
 marshmallow==3.8.0        # via -r requirements.txt, marshmallow-dataclass, marshmallow-enum, marshmallow-polyfield
