+++ requirements/requirements.in
@@ -10,7 +10,7 @@ Flask-Cors
 Flask-RESTful
 gevent>=1.5a3
 guppy3
-marshmallow-dataclass==6.0.0
+marshmallow-dataclass
 marshmallow-polyfield
 marshmallow
 marshmallow_enum
