+++ requirements/requirements.txt
@@ -42,7 +42,7 @@ jinja2==2.10.1            # via flask
 jsonschema==3.2.0         # via web3
 lru-dict==1.1.6           # via web3
 markupsafe==1.1.1         # via jinja2
-marshmallow-dataclass==6.0.0  # via -r requirements.in
+marshmallow-dataclass==8.0.0  # via -r requirements.in
 marshmallow-enum==1.5.1   # via -r requirements.in
 marshmallow-polyfield==5.9  # via -r requirements.in
 marshmallow==3.8.0        # via -r requirements.in, marshmallow-dataclass, marshmallow-enum, marshmallow-polyfield
